Give api models more intuitive names
